Move page lock/unlock views to wagtail.admin.views.pages.lock
